requirements.txt
Flask>=2.1
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.1
Flask-Mail>=0.9.1
python-dotenv>=1.0
Pillow>=9.0
gunicorn>=20.1.0
